+++ src/main/java/seedu/address/model/components/tag/Tag.java
@@ -1,4 +1,4 @@
-package seedu.address.model.tag;
+package seedu.address.model.components.tag;
 
 import static java.util.Objects.requireNonNull;
 import static seedu.address.commons.util.AppUtil.checkArgument;
