+++ src/main/java/seedu/address/model/util/SampleDataUtil.java
@@ -7,10 +7,10 @@ import java.util.stream.Collectors;
 import seedu.address.model.ReadOnlyTutorsPet;
 import seedu.address.model.TutorsPet;
 import seedu.address.model.components.name.Name;
+import seedu.address.model.components.tag.Tag;
 import seedu.address.model.student.Email;
 import seedu.address.model.student.Student;
 import seedu.address.model.student.Telegram;
-import seedu.address.model.tag.Tag;
 
 /**
  * Contains utility methods for populating {@code TutorsPet} with sample data.
