+++ src/main/java/seedu/address/ui/StudentCard.java
@@ -11,8 +11,8 @@ import javafx.scene.control.Label;
 import javafx.scene.layout.FlowPane;
 import javafx.scene.layout.HBox;
 import javafx.scene.layout.Region;
+import seedu.address.model.components.tag.Tag;
 import seedu.address.model.student.Student;
-import seedu.address.model.tag.Tag;
 
 /**
  * A UI component that displays information of a {@code Student}.
