+++ src/test/java/seedu/address/testutil/StudentBuilder.java
@@ -5,10 +5,10 @@ import java.util.Set;
 import java.util.UUID;
 
 import seedu.address.model.components.name.Name;
+import seedu.address.model.components.tag.Tag;
 import seedu.address.model.student.Email;
 import seedu.address.model.student.Student;
 import seedu.address.model.student.Telegram;
-import seedu.address.model.tag.Tag;
 import seedu.address.model.util.SampleDataUtil;
 
 /**
